OneDrive/Desktop/project/pyproject.toml
[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.env
    | _build
    | postgres_data
    | buck-out
    | build
    | dist
  )/
)
'''


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120


[tool.poetry]
name = "fastapi-project"
version = "0.1.0"
description = ""
authors = ["Бекарюков Максим Викторович <mvbekaryuk@mts.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
pydantic = "^2.6.0"
icecream = "^2.1.3"
orjson = "^3.9.12"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-bugbear = "^24.1.17"
pytest = "^8.0.0"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
